pyproject.toml
[tool.poetry]
name = "cybulde"
version = "0.1.0"
description = ""
authors = ["Sacha van Weeren <sacha.van.weeren@gmail.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
hydra-core = "~=1.3.2"
pydantic = "~=1.10.7"
pandas = "~=2.0"
#numpy = "~=2.1.2"
nltk = "~=3.8"
symspellpy = "~=6.7"
fsspec = {version = "^2024.9", extras=["gcs", "s3"]}
gcsfs = "^2024.9"
google_cloud_secret_manager = "~=2.16"
fastparquet = "~=2024.2"
dvc = {version = "^3.5", extras = ["gdrive", "gs", "s3", "dvc-s3"]}
#dask = {version = "~=2024.9.1", extras=["complete"]}
dask-expr = "~=1.1.15"
dask = {version = "*", extras=["complete"]}
#distributed = "~=2024.9"
#llvmlite = "~=0.43"
distributed = "*"
#dask-ml = "~=2023.3"
dask-ml = "*"
dask-cloudprovider = {version = "2024.9", extras =["gcp","aws"]}
awscli = "^1"
boto3 = "*"
jupyterlab = "*"

[tool.poetry.group.dev.dependencies]
pytest = "~=7.3"
black = "~=23.3"
isort = "~=5.12"
flake8 = "~=6.0"
mypy = "~=1.2"
jupyterlab = "~=4.2"
types-PyYAML = "~=6.0.12.10" 
types-decorator = "~=5.1.8.3" 
types-paramiko = "~=3.0.0.10" 
types-protobuf = "~=4.23.0.1" 
types-pyOpenSSL = "~=23.1.0.3" 
types-python-dateutil = "~=2.8.19.13" 
types-requests = "~=2.30.0.0" 
types-setuptools = "~=67.8.0.0" 
types-six = "~=1.16.21.8" 
types-typed-ast = "~=1.5.8.6" 
types-urllib3 = "~=1.26.25.13"
types-pycurl= "~=7.45.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["cybulde"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
